+++ Docs/Changes.md
@@ -8,6 +8,7 @@
   * Fix: _Dya is unable to find config file_ ([Issue #943](https://github.com/vorov2/dyalect/issues/943)).
   * Fix: _Lazy variables don't work correctly in interactive_ ([Issue #944](https://github.com/vorov2/dyalect/issues/944)).
   * Fix: _Unary `+` is not compiled_ ([Issue #946](https://github.com/vorov2/dyalect/issues/946)).
+  * Fix: _Disallow `final` and `abstract` modifiers on the same function_ ([Issue #947](https://github.com/vorov2/dyalect/issues/947)).
 
 # 0.46.7
   * Fix: _Linker may crush in a case of incorrect module import_ ([Issue #939](https://github.com/vorov2/dyalect/issues/939)).